Adding Basic Scanner and Hash Files
Scanner is currently able to scan all files on a windows computer, check against a list of known malware hashes, and delete detected files. 
Hashes file consists of 1k+ real malicious files, plus a test file as the last hash.
testfile.txt is the test file that is stored as the lash hash in the hashes file.